Add storyboards {scenario 1, 2 and 3}
Storyboarding is used in HCI and SDD as part of identifying the specifications for a particular software/application. It is useful as it helps the user understand exactly how the our HTR application will work, much better than an abstract description. It is also cheaper as we can make changes to a storyboard with minimal costs than an implemented piece of software through prototyping.